backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==3.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.122.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
fastapi-mail
ccxt  # This includes ccxt.pro automatically
pandas
numpy
backtrader
google-genai
quantstats
matplotlib
seaborn
IPython
TA-Lib
python-dotenv
celery==5.3.6
redis==5.0.1
alembic
deap
tqdm
alpaca-py
dotenv
pandas_ta
asgiref>=3.7.0
weasyprint>=60.0
textblob
praw==7.7.1
vaderSentiment==3.3.2
feedparser==6.0.10
gnews
httpx
feedparser==6.0.10
beautifulsoup4
requests==2.32.5
lxml==4.9.3
feedparser
python-telegram-bot
transformers
torch
scipy
deep-translator
python-dateutil
spacy
aiohttp
statsmodels
hmmlearn
yfinance
docker>=7.0.0  # Docker SDK for log monitoring